+++ src/menu.h
@@ -2,7 +2,7 @@
 
 #include "config.h"
 
-#define MENU_SIZE 115
+#define MENU_SIZE 116
 
 
 MENU_ITEM menuItemsSource[MENU_SIZE] = {
@@ -84,13 +84,14 @@ MENU_ITEM menuItemsSource[MENU_SIZE] = {
     {MENU_SDCARD_AUTOSTARTLOG,      MENU_SDCARD,        MENU_NO_MENU,     "Autostart log enabled"},
     {MENU_SDCARD_MOUNT_STATUS,      MENU_SDCARD,        MENU_NO_MENU,     "Status"},
     {MENU_SDCARD_REC,               MENU_SDCARD,        MENU_NO_MENU,     "Record"},
-    {MENU_SDCARD_SETTINGS_SAVE,     MENU_SDCARD,        MENU_NO_MENU,     "Save settings to sd"},
-    {MENU_SDCARD_SETTINGS_RESTORE,  MENU_SDCARD,        MENU_NO_MENU,     "Restore settings from sd"},
+    {MENU_SDCARD_SETTINGS_SAVE,     MENU_SDCARD,        MENU_NO_MENU,     "Backup settings to SDCARD"},
+    {MENU_SDCARD_SETTINGS_RESTORE,  MENU_SDCARD,        MENU_NO_MENU,     "Restore settings from SD"},
 //  {MENU_SDCARD_INTERVAL,          MENU_SDCARD,        MENU_NO_MENU,     "Log interval sec."},
 
     {MENU_VOLTMETER_TOP,            MENU_VOLTMETER,     MENU_OTHERS,      "<- parent menu"},
     {MENU_VOLTMETER_ENABLED,        MENU_VOLTMETER,     MENU_NO_MENU,     "Voltmeter enabled"},
     {MENU_VOLTMETER_SLEEP,          MENU_VOLTMETER,     MENU_NO_MENU,     "Control SleepMode"},
+    {MENU_VOLTMETER_INFO,           MENU_VOLTMETER,     MENU_NO_MENU,     "Show voltage/current"},
     {MENU_VOLTMETER_SLEEPVOL,       MENU_VOLTMETER,     MENU_NO_MENU,     "Sleep Vol."},
     {MENU_VOLTMETER_WAKEUPVOL,      MENU_VOLTMETER,     MENU_NO_MENU,     "WakeUp Vol."},
     {MENU_VOLTMETER_CUTOFFVOL,      MENU_VOLTMETER,     MENU_NO_MENU,     "CutOff Vol."},